Spread scan tasks over Ray cluster.
